# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Documentation for the Orders application",
        "contact": {
            "email": "moazsamy333@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/users/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user in the database",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Failure"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login as user",
                "description": "Logs in as a user",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unmatching credentials"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/logout": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout",
                "description": "Logout from the system",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Not logged in"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginUserRequest": {
                "title": "Login User request",
                "description": "Login User request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "User email",
                        "description": "User email",
                        "type": "string",
                        "example": "moazsamy333@gmail.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Username",
                        "type": "string",
                        "example": "Moaz"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User email",
                        "type": "string",
                        "example": "moazsamy333@gmail.com"
                    },
                    "type": {
                        "title": "User Type",
                        "description": "User type. 0-Admin, 1-Delivery Personnel",
                        "type": "integer",
                        "example": "1"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "RegisterUserRequest": {
                "title": "Register User request",
                "description": "Register User request body data",
                "required": [
                    "email",
                    "password",
                    ""
                ],
                "properties": {
                    "name": {
                        "title": "User name",
                        "description": "User name",
                        "type": "string",
                        "example": "Moaz"
                    },
                    "email": {
                        "title": "User email",
                        "description": "User email",
                        "type": "string",
                        "example": "moazsamy333@gmail.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in from (bearer token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}